+++ test/unit/test_backbone_global_net.py
@@ -103,6 +103,8 @@ class TestGlobalNet:
             concat_skip=False,
             encode_kernel_sizes=[7, 3, 3],
             decode_kernel_sizes=3,
+            encode_num_channels=[2, 4, 8],
+            decode_num_channels=[2, 4, 8],
             strides=2,
             padding="same",
             name="Test",