+++ CHANGELOG.md
@@ -13,6 +13,7 @@ compatible with the updates.
 
 ### Added
 
+- Added support to custom layer channels in U-Net.
 - Added support to multiple loss functions for each loss type: "image", "label" and
   "regularization".
 - Added LNCC computation using separable 1-D filters for all kernels available